contacts-manager/contacts-manager-corrige/src/main/java/com/fredericboisguerin/insa/contactsmanager/model/Contact.java
package com.fredericboisguerin.insa.contactsmanager.model;

public class Contact {
    private String name;
    private String email;
    private String phoneNumber;

    public Contact(String name, String email, String phoneNumber) {
        this.name = name;
        this.email = email;
        this.phoneNumber = phoneNumber;
    }

    public String getName() {
        return name;
    }

    public String getEmail() {
        return email;
    }

    public String getPhoneNumber() {
        return phoneNumber;
    }

    public boolean nameMatches(String pattern) {
        return this.name.toLowerCase().contains(pattern);
    }
}
